refactor: move base modules from unilab/envs to unilab/base
- Move env_base.py, np_env.py, registry.py, curriculum.py, dtype_config.py to unilab/base/
- Move backend/ directory to unilab/base/
- Update all imports across locomotion and manipulation modules
- Add backward compatibility re-exports in unilab/envs/__init__.py
