+++ unilab/envs/__init__.py
@@ -1,1 +1,6 @@
 """Environment registry and base classes."""
+
+# Re-export from unilab.base for backward compatibility
+from unilab.base import EnvCfg, NpEnv, registry, EpisodeLengthTracker, PenaltyCurriculum
+
+__all__ = ["EnvCfg", "NpEnv", "registry", "EpisodeLengthTracker", "PenaltyCurriculum"]
