+++ unilab/envs/locomotion/g1/base.py
@@ -5,9 +5,9 @@ import mujoco
 import numpy as np
 from dataclasses import dataclass, field
 
-from unilab.envs.base import EnvCfg
-from unilab.envs.np_env import NpEnv, NpEnvState
-from unilab.envs.backend import SimBackend
+from unilab.base.env_base import EnvCfg
+from unilab.base.np_env import NpEnv, NpEnvState
+from unilab.base.backend import SimBackend
 
 
 @dataclass
